requirements.txt
alembic==1.16.5

Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-login
gunicorn==23.0.0

Jinja2==3.1.6
Flask-WTF==1.2.1
numpy==2.3.2
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.1.0
ply==3.11
productmd==1.45
psycopg==3.2.9
psycopg2-binary==2.9.10
python-slugify==8.0.4
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
qrcode==8.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
python-dotenv